backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
pydantic-settings
llama-index
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
chromadb
sentence-transformers